GH-2970: minor improvement error handler related
Fixes: #2970 

* minor improvement error handler related
* remove `FailedRecordProcessor.retryListeners`, reuse FailureTracker's retryListeners
* cleanup related to error handlers

+++ spring-kafka/src/main/java/org/springframework/kafka/listener/FallbackBatchErrorHandler.java
@@ -181,7 +181,7 @@ class FallbackBatchErrorHandler extends ExceptionClassifier implements CommonErr
 					.stream()
 					.collect(
 							Collectors.toMap(tp -> tp,
-									tp -> data.records(tp).get(0).offset(), (u, v) -> (long) v, LinkedHashMap::new))
+									tp -> data.records(tp).get(0).offset(), (u, v) -> v, LinkedHashMap::new))
 					.forEach(consumer::seek);
 
 			throw new KafkaException("Seek to current after exception", getLogLevel(), thrownException);
